+++ airbrake/handler.go
@@ -8,9 +8,13 @@ Create a new handler as you would create a gobrake.Notifier:
 
 	handler := airbrake.NewHandler(projectID, projectKey)
 
-If you need access to the underlying Notifier instance (or need more advanced construction), you can access it from the handler:
+If you need access to the underlying Notifier instance (or need more advanced construction), you can create the handler from a notifier directly:
 
-	handler.Notifier.SetHost("https://errbit.domain.com")
+	projectID  := int64(1)
+	projectKey := "key"
+
+	notifier := gobrake.NewNotifier(projectID, projectKey)
+	handler := airbrake.NewHandlerFromNotifier(notifier)
 
 By default Gobrake sends errors asynchronously and expects to be closed before the program finishes:
 
@@ -20,7 +24,7 @@ By default Gobrake sends errors asynchronously and expects to be closed before t
 
 If you want to Flush notices you can do it as you would with Gobrake's notifier or you can configure the handler to send notices synchronously:
 
-	handler.SendSynchronously = true
+	handler := airbrake.NewHandlerFromNotifier(notifier, airbrake.SendSynchronously(true))
 */
 package airbrake
 
@@ -31,42 +35,74 @@ import (
 	"github.com/goph/emperror/internal/keyvals"
 )
 
-// Handler is responsible for sending errors to Airbrake/Errbit.
-type Handler struct {
-	Notifier          *gobrake.Notifier
-	SendSynchronously bool
+// Option configures a logger instance.
+type Option interface {
+	apply(*handler)
+}
+
+// SendSynchronously configures the handler to send notices synchronously.
+type SendSynchronously bool
+
+func (o SendSynchronously) apply(l *handler) {
+	l.sendSynchronously = bool(o)
+}
+
+// handler is responsible for sending errors to Airbrake/Errbit.
+type handler struct {
+	notifier *gobrake.Notifier
+
+	sendSynchronously bool
 }
 
 // NewHandler creates a new Airbrake handler.
-func NewHandler(projectID int64, projectKey string) *Handler {
-	return &Handler{
-		Notifier: gobrake.NewNotifier(projectID, projectKey),
+func NewHandler(projectID int64, projectKey string, opts ...Option) *handler {
+	h := &handler{
+		notifier: gobrake.NewNotifier(projectID, projectKey),
 	}
+
+	for _, o := range opts {
+		o.apply(h)
+	}
+
+	return h
+}
+
+// NewHandlerFromNotifier creates a new Airbrake handler from a notifier instance.
+func NewHandlerFromNotifier(notifier *gobrake.Notifier, opts ...Option) *handler {
+	h := &handler{
+		notifier: notifier,
+	}
+
+	for _, o := range opts {
+		o.apply(h)
+	}
+
+	return h
 }
 
 // Handle calls the underlying Airbrake notifier.
-func (h *Handler) Handle(err error) {
+func (h *handler) Handle(err error) {
 	// Get HTTP request (if any)
 	req, _ := httperr.HTTPRequest(err)
 
 	// Expose the stackTracer interface on the outer error (if there is stack trace in the error)
 	err = emperror.ExposeStackTrace(err)
 
-	notice := h.Notifier.Notice(err, req, 1)
+	notice := h.notifier.Notice(err, req, 1)
 
 	// Extract context from the error and attach it to the notice
 	if kvs := emperror.Context(err); len(kvs) > 0 {
 		notice.Params = keyvals.ToMap(kvs)
 	}
 
-	if h.SendSynchronously {
-		h.Notifier.SendNotice(notice)
+	if h.sendSynchronously {
+		h.notifier.SendNotice(notice)
 	} else {
-		h.Notifier.SendNoticeAsync(notice)
+		h.notifier.SendNoticeAsync(notice)
 	}
 }
 
 // Close closes the underlying Airbrake instance.
-func (h *Handler) Close() error {
-	return h.Notifier.Close()
+func (h *handler) Close() error {
+	return h.notifier.Close()
 }