+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **errorlog:** Renamed log package
 - **errorlog:** Improved package
 - **errorlogrus:** Rename logrus package
+- **airbrake:** Improve Airbrake package
 
 
 ## [0.9.1] - 2018-07-27